+++ internal/runner/runner.go
@@ -8,7 +8,6 @@ import (
 	_ "net/http/pprof"
 	"os"
 	"path/filepath"
-	"reflect"
 	"strings"
 	"sync/atomic"
 	"time"
@@ -68,6 +67,11 @@ var (
 	// EnableCloudUpload is global variable to enable cloud upload
 	EnableCloudUpload = false
 )
+
+func init() {
+	HideAutoSaveMsg = env.GetEnvOrDefault("DISABLE_CLOUD_UPLOAD_WRN", false)
+	EnableCloudUpload = env.GetEnvOrDefault("ENABLE_CLOUD_UPLOAD", false)
+}
 
 // Runner is a client for running the enumeration process.
 type Runner struct {
@@ -713,58 +717,4 @@ func (r *Runner) SaveResumeConfig(path string) error {
 	data, _ := json.MarshalIndent(resumeCfgClone, "", "\t")
 
 	return os.WriteFile(path, data, permissionutil.ConfigFilePermission)
-}
-
-type WalkFunc func(reflect.Value, reflect.StructField)
-
-// Walk traverses a struct and executes a callback function on each value in the struct.
-// The interface{} passed to the function should be a pointer to a struct or a struct.
-// WalkFunc is the callback function used for each value in the struct. It is passed the
-// reflect.Value and reflect.Type properties of the value in the struct.
-func Walk(s interface{}, callback WalkFunc) {
-	structValue := reflect.ValueOf(s)
-	if structValue.Kind() == reflect.Ptr {
-		structValue = structValue.Elem()
-	}
-	if structValue.Kind() != reflect.Struct {
-		return
-	}
-	for i := 0; i < structValue.NumField(); i++ {
-		field := structValue.Field(i)
-		fieldType := structValue.Type().Field(i)
-		if !fieldType.IsExported() {
-			continue
-		}
-		if field.Kind() == reflect.Struct {
-			Walk(field.Addr().Interface(), callback)
-		} else if field.Kind() == reflect.Ptr && field.Elem().Kind() == reflect.Struct {
-			Walk(field.Interface(), callback)
-		} else {
-			callback(field, fieldType)
-		}
-	}
-}
-
-// expandEndVars looks for values in a struct tagged with "yaml" and checks if they are prefixed with '$'.
-// If they are, it will try to retrieve the value from the environment and if it exists, it will set the
-// value of the field to that of the environment variable.
-func expandEndVars(f reflect.Value, fieldType reflect.StructField) {
-	if _, ok := fieldType.Tag.Lookup("yaml"); !ok {
-		return
-	}
-	if f.Kind() == reflect.String {
-		str := f.String()
-		if strings.HasPrefix(str, "$") {
-			env := strings.TrimPrefix(str, "$")
-			retrievedEnv := os.Getenv(env)
-			if retrievedEnv != "" {
-				f.SetString(os.Getenv(env))
-			}
-		}
-	}
-}
-
-func init() {
-	HideAutoSaveMsg = env.GetEnvOrDefault("DISABLE_CLOUD_UPLOAD_WRN", false)
-	EnableCloudUpload = env.GetEnvOrDefault("ENABLE_CLOUD_UPLOAD", false)
 }
